[looper] Improved graph generation
Graph generated for XNU example now does not
contain some useless join nodes. It is by no means
complete, need to test it with more examples.
One thing that needs to be done is aggregation
of multiple nodes that are created when prune
condition is complex and contains logical operators.

Started working on mechanism for assembling of
LTS edge formulas which now has priority.