[looper] Output of basic DCP finished
First version of DCP generating algorithm is
finished. Still missing guards and second
abstraction step: from integers to natural numbers.

There's also a problem with local variable scopes
as I mentioned in previous commits. Top level scope
of all local variables leads to generation of
invalid norms and consequently to generation
of invalid difference constraints on edges. Don't
know how to fix this yet and it might not even be
a problem for the bound algorithm, we'll see
once we get there.