[looper] Basic transition guard deriving
First and incomplete version of guard deriving
algorithm is finished. It does not support all
possible variants of condition expressions but
it can generate guards for simple ones. It will
be extended later when we'll support for rest of
the cases cases.

Guard propagation needs to be implemented next
and I must also figure out a way how to prevent
generation of unwanted norms. It most likely wont
be perfect but we are currently limited by SIL/CFG
which does not provide any info about location
of variable declarations.