[looper] Correct DCP generation
Fixed generation of invalid difference constraints.
Added check if all variables in a norm are locals
that are defined at certain location or formals.
DC for specified norm is not generated if this
condition does not hold.

Also rewrote part of the DC derivation algorithm
to make it faster and more clear.