[looper] Checking for bound variable monotonicity with Z3
Groundwork for more precise inter-procedural
analysis. Trying to check monotonicity
(increasing/decreasing) of each bound variable
(function parameters) with Z3. We basically need
to determine if we should look for minimum
or maximum for each argument when instantiating
function summary so that we obtain sound bound.
However, we need to use universal quantifiers
for these checks and on top of that the expressions
are frequently non-linear so these checks are
undecidable. In my limited testing Z3 indeed seems
to fail/loop on some expressions so we
will need to maybe think of some better way to
solve this / use heuristics