[looper] Improved monotony checking + restructuring
Checks for bound variable monotonicity with
Z3 were reworked a little. Switched to Real
type sorts because integer arithmetic caused
timeout issues / unknown results. Real arithmetic
seems to work fine and the results proven over
reals should be applicable to integers in our case
(I think?). Also, the function body in the quantified expression is now constructed
for each bound parameter. This allows for a
single bound variable in the quantified expression.
I'm not quite sure if it is correct this way (must
check later), but it's faster and the results seem
alright.

Also restructured EdgeExp type a little in order
to decouple it with summary type which was
needed for further work.