[looper] SIL to EdgeExp conversion with pointers
Trying to fix the issue with conversion of SIL expressions to EdgeExp so that we can also support
pointers and dereferencing which is necessary if we
want to track side-effects of functions. However the
conversion is not working correctly sometimes and
this change might have broken something that was
previously working.