[looper] Side-effect tracking
It is now possible to compute variable bounds for
values under formals of pointer types. Still work in
progress but it seems that it might be possible to
do it this way. Some hack-around solutions were
necessary because we cannot treat formals as
symbolic constants and variables at the same time
so some trees in reset graphs don't have symbolic
constants roots.