[Refactoring, graph optimizations]
Refactored old code and improved the graph generation
algorithm to merge multiple Prune nodes together if they belong
to the same loophead. Conditions are also accumulated.
This was necessary in order to have all loop conditions on
single DCP edge which will form basis for further precision
improvement to the bound algorithm.