[Compound norm analysis, reporting, exceptions]
Changed the local bound algorithm to respect the original
conditions in the source code and to work with compound norms.
This leads to a more precise analysis where:
x > 0 && y > 0 leads to min(tb(x), tb(y)) and
x > 0 || y > 0 leads to max(tb(x), tb(y)).

Started figuring out how to report issues in Infer properly
and added excpetion handling so that Infer does not crash on
issues